add: insert product with typeorm

+++ src/products/entities/product.entity.ts
@@ -8,7 +8,7 @@ export class Product {
   id: string;
   @Column('text', { unique: true })
   title: string;
-  @Column('numeric', {
+  @Column('float', {
     default: 0,
   })
   price: number;
